+++ apps/palette/CHANGELOG.md
@@ -1,5 +1,35 @@
 # palette
 
+## 1.4.41
+
+### Patch Changes
+
+- @saas-ui/react@2.8.9
+- @saas-ui/theme-glass@0.5.5
+
+## 1.4.40
+
+### Patch Changes
+
+- @saas-ui/react@2.8.8
+- @saas-ui/theme-glass@0.5.5
+
+## 1.4.39
+
+### Patch Changes
+
+- @saas-ui/react@2.8.7
+- @saas-ui/theme-glass@0.5.5
+
+## 1.4.38
+
+### Patch Changes
+
+- Updated dependencies [12c35ea]
+- Updated dependencies [b61ff73]
+  - @saas-ui/react@2.8.6
+  - @saas-ui/theme-glass@0.5.5
+
 ## 1.4.37
 
 ### Patch Changes