+++ packages/saas-ui-auth0/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @saas-ui/auth0
 
+## 0.1.11
+
+### Patch Changes
+
+- @saas-ui/auth@3.2.1
+
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [dadcd94]
+  - @saas-ui/auth@3.2.0
+
+## 0.1.9
+
+### Patch Changes
+
+- @saas-ui/auth@3.1.9
+
+## 0.1.8
+
+### Patch Changes
+
+- @saas-ui/auth@3.1.8
+
 ## 0.1.7
 
 ### Patch Changes