+++ packages/saas-ui-modals/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @saas-ui/modals
 
+## 2.1.39
+
+### Patch Changes
+
+- Updated dependencies [df5d37f]
+  - @saas-ui/forms@2.6.9
+
+## 2.1.38
+
+### Patch Changes
+
+- Updated dependencies [fc62212]
+  - @saas-ui/forms@2.6.8
+
+## 2.1.37
+
+### Patch Changes
+
+- 02cac87: Updated dependencies
+- Updated dependencies [02cac87]
+  - @saas-ui/forms@2.6.7
+
+## 2.1.36
+
+### Patch Changes
+
+- @saas-ui/forms@2.6.6
+
 ## 2.1.35
 
 ### Patch Changes